requirements.txt
google-api-python-client>=2.106.0
gspread>=5.7.1
nltk>=3.8.1
openai>=0.27.4
pytest>=7.4.0
python-decouple>=3.8
Requests>=2.31.0
scikit-learn>=1.4.1.post1
selenium>=3.141.0
splinter>=0.21.0
torch>=2.2.1
transformers>=4.39.0
